upgrade project to api 28 user support Fragment when possible as the system fragment is deprecated and doesn't work well with the new support library anymore.

+++ library/src/main/java/net/alhazmy13/hijridatepicker/date/hijri/MonthView.java
@@ -506,8 +506,8 @@ public abstract class MonthView extends View {
 
             final int startX = (int) (x - dayWidthHalf);
             final int stopX = (int) (x + dayWidthHalf);
-            final int startY = (int) (y - yRelativeToDay);
-            final int stopY = (int) (startY + mRowHeight);
+            final int startY = (y - yRelativeToDay);
+            final int stopY = (startY + mRowHeight);
 
             drawMonthDay(canvas, mYear, mMonth, dayNumber, x, y, startX, stopX, startY, stopY);
